Backend: Ensure validation of `habitid` and `recordid` to prevent invalid UUIDs from causing errors or unintended behavior

+++ HabitVault-backend/routes/api/user/habits.ts
@@ -3,7 +3,7 @@ import { PrismaClient } from '@prisma/client'
 
 import { restrict } from '../../../middlewares'
 import { TypedRequest, HabitBody, HabitRecordBody } from '../../../interfaces'
-import { isValidHabitName, calculateStreak, calculateMaxStreak } from '../../../utils'
+import { isValidHabitName, calculateStreak, calculateMaxStreak, isValidUUIDV4 } from '../../../utils'
 
 const router = Router()
 const prisma = new PrismaClient()
@@ -77,7 +77,7 @@ router.post('/', restrict, async (req: TypedRequest<HabitBody>, res: Response) =
 router.get('/:id', restrict, async (req: TypedRequest<any, { id: string }>, res: Response) => {
     const userid = req.session.username
     const habitid = req.params.id
-    if (userid !== undefined && habitid !== undefined) {
+    if (userid !== undefined && habitid !== undefined && isValidUUIDV4(habitid)) {
         const habit = await prisma.habit.findFirst({
             where: {
                 id: habitid,
@@ -96,7 +96,7 @@ router.put('/:id', restrict, async (req: TypedRequest<HabitBody, { id: string }>
     const habitid = req.params.id
     const name = req.body.name
     const description = req.body.description
-    if (userid !== undefined && name !== undefined && description !== undefined && habitid !== undefined) {
+    if (userid !== undefined && name !== undefined && description !== undefined && habitid !== undefined && isValidUUIDV4(habitid)) {
         const habit = await prisma.habit.update({
             where: { id: habitid },
             data: {
@@ -115,7 +115,7 @@ router.put('/:id', restrict, async (req: TypedRequest<HabitBody, { id: string }>
 router.delete('/:id', restrict, async (req: TypedRequest<HabitBody, { id: string }>, res: Response) => {
     const habitid = req.params.id
     const userid = req.session.username
-    if (userid !== undefined && habitid !== undefined) {
+    if (userid !== undefined && habitid !== undefined && isValidUUIDV4(habitid)) {
         const habit = await prisma.habit.delete({
             where: {
                 id: habitid,
@@ -131,23 +131,33 @@ router.delete('/:id', restrict, async (req: TypedRequest<HabitBody, { id: string
 router.get('/:id/streak', restrict, async (req: TypedRequest<any, { id: string }>, res: Response) => {
     const userid = req.session.username
     const habitid = req.params.id
-    const dates = await prisma.habitRecord.findMany({
-        where: { habitId: habitid, userId: userid },
-        select: { date: true },
-        orderBy: { date: 'desc' }
-    })
-    const d = dates.map(d => d.date)
-    res.json({ streak: calculateStreak(d), max_streak: calculateMaxStreak(d) })
+    if (userid !== undefined && habitid !== undefined && isValidUUIDV4(habitid)) {
+        const dates = await prisma.habitRecord.findMany({
+            where: { habitId: habitid, userId: userid },
+            select: { date: true },
+            orderBy: { date: 'desc' }
+        })
+        const d = dates.map(d => d.date)
+        res.json({ streak: calculateStreak(d), max_streak: calculateMaxStreak(d) })
+    }
+    else {
+        res.status(400).json()
+    }
 })
 
 router.get('/:id/records', restrict, async (req: TypedRequest<any, { id: string }>, res: Response) => {
     const userid = req.session.username
     const habitid = req.params.id
-    const habit = await prisma.habit.findFirst({
-        where: { id: habitid, userId: userid },
-        include: { records: true }
-    })
-    res.json(habit?.records)
+    if (userid !== undefined && habitid !== undefined && isValidUUIDV4(habitid)) {
+        const habit = await prisma.habit.findFirst({
+            where: { id: habitid, userId: userid },
+            include: { records: true }
+        })
+        res.json(habit?.records)
+    }
+    else {
+        res.status(400).json()
+    }
 })
 
 //TODO: data validation, check if record already exists 
@@ -155,7 +165,7 @@ router.post('/:id/records/', restrict, async (req: TypedRequest<HabitRecordBody,
     const userid = req.session.username
     const habitid = req.params.id
     const date = req.body.date
-    if (date !== undefined && userid !== undefined && habitid !== undefined) {
+    if (date !== undefined && userid !== undefined && habitid !== undefined && isValidUUIDV4(habitid)) {
         const habit = await prisma.habitRecord.create({
             data: {
                 habitId: habitid,
@@ -176,7 +186,7 @@ router.delete('/:id/records/:recordid', restrict, async (req: TypedRequest<any, 
     const recordid = req.params.recordid
 
 
-    if (userid !== undefined && habitid !== undefined && recordid !== undefined) {
+    if (userid !== undefined && habitid !== undefined && recordid !== undefined && isValidUUIDV4(habitid) && isValidUUIDV4(recordid)) {
         const habit = await prisma.habitRecord.delete({
             where: { id: recordid, habitId: habitid, userId: userid },
         })
